Fix guess command ran decoders plugin for both sampling and guess
When running sampling file in SamplingParserPlugin.runFileInputSampling,
input config.yml is used. The method is called from
FileInputRunner.guess() and FileInputRunner.guess() calls
guessParserConfig() for the sampling buffer. In guessParserConfig(),
FileInputRunner runs decoders again to guess a file format.

If 'decoders' is set, it is run for both sampling and guess. It will
cause a problem when setting decoders in config.yml for guess command.