Refactor to alloc visitor instances in while loop
Commonly, it may be high cost to create instances in a loop. Current
CsvFormatterPlugin and CsvParserPlugin creates ColumnVisitor for each
record. It may be better to create a new instance outside of the loop.

When using hprof profile for heap=site in my test and it shows 0.04%
accumulate.(It depends on the number of records.) So, it affects very
few memory allocation and it may be difficult to measure the performance
improvement.