CSV guess plugin uses variance of column lengths to detect heder line
This change enhances CSV guess plugin so that it may detect header line
even if all values are string. This guess succeeds if values in a CSV
file are (almost) fixed-length and its column name at the first line is
longer or shorter enough compared to length of values. Fixes #284.