Support "mkbundle --path" option
When we deploy embulk to a production environment using chef or
capistrano, gems should be installed to a shared directory
(shared/bundler) but Gemfile.lock should be isolated. --path optoin
works for that case.

This change includes bundler itself to embulk-core.jar so that
embulk_run doesn't have to install bundler. Thus initialization of
bundler becomes less magical.