LocalExecutorPlugin runs more output threads by scattering input pages
This change implements #232.

LocalExecutorPlugin creates N times more output tasks and run them in
parallel if number of input tasks is less than min_threads option
(N > 1). Default min_threads is same with number of CPU cores.

This behavior is deterministic as long as min_threads option is not
changed.