Avoid IllegalArgumentException in try-with-resource
When using try-with-resource block and throw Exception in both the block
and close() method, 'IllegalArgumentException: Self-suppression not
permitted' is thrown. LocalExecutorPlugin throws the exception when
OutputPlugin throws RuntimeException. RuntimeException is caught by
ExecutorService and the result is returned via Future object in
OutputWorkers. When OutputPlugin throws RuntimeException, Future.get()
throws ExecutionException. The exception is handled by
ScatterTransactionalPageOutput.completeWorkers(). completeWorkers()
are called by finish(), close(), commit(), and abort().
When finish()(or commit()) throws RuntimeException via completeWorkers(),
close()(and abort() via AbortTransactionResource) also throws
RuntimeException. From this, IAE is thrown.

This fix is to store a first RuntimeException as reportedException and
then check it in close() or abort() to avoid throwing the same exception
again.