Add dontAbort() after commit() to skip abort()
I think dontAbort() should be called after commit() call
when using AbortTransactionResource.
AbortTransactionResource is used in try-with-resource block.
When leaving the block, tran.abort() is called. When commit()
is finished without failure, the abort() call is not required.
From other usage of AbortTransactionResource in embulk,
dontAbort() is called after commit() or just before leaving
try-with-resource block. I think runInputTask() should call
dontAbort() at the end of the block.