Enhance JsonParserPlugin * Add file_type and object_field options. The plugin can parse several types of JSON files. currently it supports [sequence, object, array, lines] as file types. sequence is by default.
If `field_type: sequence` specified, the plugin can parse the sequence of JSON objects as records:

```
{"col1":"val1","col2":"val2"}{"col1":"val3","col2":"val4"}{"col1":"val1","col5":"val6"}{"col1":"val7","col2":"val8"}
```
and
```
{
  "col1":"val1",
  "col2":"val2"
}
{
  "col1":"val3",
  "col2":"val4"
}
{
  "col1":"val1",
  "col5":"val6"
}
{
  "col1":"val7",
  "col2":"val8"
}
```

By `field_type: object`, it can extract and parse the value in a big JSON object in the file as records:

```
{"records":[
  {"col1":"val1","col2":"val2"},
  {"col1":"val3","col2":"val4"},
  {"col1":"val1","col5":"val6"},
  {"col1":"val7","col2":"val8"}
]}
```

The extracted key is `records` by default. Users can specify the key name as object_field option.

`field_type: array` can parse the following a big JSON array in the file.

```
[
  {"col1":"val1","col2":"val2"},
  {"col1":"val3","col2":"val4"},
  {"col1":"val1","col5":"val6"},
  {"col1":"val7","col2":"val8"}
]
```

`field_type: lines` can parse a JSON file in line-json format.