Update the repo for https://www.embulk.org/docs from "www.embulk.org" to "docs"
We had had maintained the entire www.embulk.org website in the repo "www.embulk.org".
But, the repo is now renamed to "docs". This commit is to catch up with the latest status.

https://www.embulk.org/docs/ has been the main website, and it has been updated from
the "embulk-docs" directory of this "embulk" repository.

Later on, we plan to manage the website like:
* https://www.embulk.org/* (except for docs): https://github.com/embulk/embulk.github.io
* https://www.embulk.org/docs/: in https://github.com/embulk/docs