Guarantee TempFileSpace to be unique
A TempFileSpace has been created in a directory with a timestamp, such as "/tmp/2019-05-15 11:35:48.617 UTC".
But, when multiple Embulk processes (or tasks) are running, they can easily conflict.

To avoid such conflicts, this change starts to use Files#createTempDirectory to create a TempFileSpace.
Java VM guarantees that the directory created is unique.

https://docs.oracle.com/javase/8/docs/api/java/nio/file/Files.html#createTempDirectory-java.nio.file.Path-java.lang.String-java.nio.file.attribute.FileAttribute...-

The old constructor TempFileSpace(File) keeps the original behavior for compatibility,
but it is now deprecated. It should not be called from plugins, though.