Load a Maven-based plugin from a simple plugin type with Embulk system properties
It is to look for a Maven-based plugin even when only a simple plugin type is specified
like "type: ftp" in user configuration, if a corresponding Embulk system property is configured.

For example, assume a user configuration is like below:

in:
  type: ftp
  host: ...
  ...

Embulk has looked only for a RubyGem-based plugin so far in this case.

By this commit, Embulk looks for a Maven-based plugin if an Embulk system property
"plugins.input.ftp" is set like "maven:org.embulk:ftp:0.3.0".