Revive EmbulkTestRuntime
org.embulk.EmbulkTestRuntime was revised in Embulk v0.10.33 along with the Guice removal.

Pull request: https://github.com/embulk/embulk/pull/1444
Commit: 656504422c93a24581c4f45e01618dd2fadbb027

It turned out that it might not be working well after the revision.

We were not aware of that because EmbulkTestRuntime was an old obsolete test runner.
Modern plugins are tested with org.embulk.test.TestingEmbulk, the latest test runner.

This change is giving a try to revive EmbulkTestRuntime.
We may backport this to older versions of embulk-junit4.