Mix embulk-spi's classes into embulk-core
The compiled class files of "embulk-spi" are mixed into the "embulk-core" JAR file, and
the "embulk-core" Maven artifact has the dependency on "embulk-spi" as "provided" in pom.xml.

The artifacts are structured like this so that they can bypass the restriction of Java (9+)
Module System that multiple JAR files (modules) cannot contain the same Java package classes.

This mixing is not necessary when using Embulk with the executable all-in-one binary. But,
it would not work well with Java 9+ in case of using Embulk with self-hosted deployment of
the separated artifacts, embulk-spi, embulk-core, and else.